+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # FlexMessage CHANGELOG
 
+## FlexMessage 0.0.3
+
+- Renamed `Reset()` to `Flush()`
+
+**FEATURES**
+- Adding `Print()` Function
+- Implemented colors for `Print()`
+
 ## FlexMessage 0.0.2
 This new release introduces `CHANGELOG.md` :-)
-### FEATURES
+**FEATURES**
 - Reset function for `FlexMessage` to reset all notifications
 
-### BUGFIXES
+**BUGFIXES**
 - Fix for `Gosec` Workflow
 
 ## FlexMessage 0.0.1